Refacotr code; move logic that render ui to separate file
